#76 refactor: Create SocialLoginButton
