+++ src/components/auth/SocialLoginButton.tsx
@@ -1,0 +1,88 @@
+import { postOAuthUrl } from "@api/auth";
+import { HTTPSTATUS } from "@api/types";
+import googleLogo from "@assets/icons/logo/ic_google.svg";
+import kakaoLogo from "@assets/icons/logo/ic_kakao.svg";
+import naverLogo from "@assets/icons/logo/ic_naver.svg";
+import { createToast } from "@components/common/toast";
+import { WindowContext } from "@context/WindowContext";
+import designSystem from "@styles/designSystem";
+import openPopUpWindow from "@utils/openPopUpWindow";
+import { useContext } from "react";
+import styled from "styled-components";
+import { BaseSignInButton } from "./BaseSignInButton";
+
+type Props = {
+  provider: "google" | "kakao" | "naver";
+};
+
+export default function SocialLoginButton({ provider }: Props) {
+  const { onOpenPopUpWindow } = useContext(WindowContext);
+  const toast = createToast();
+
+  const onSignIn = async () => {
+    // Get Auth URL from server.
+    const res = await postOAuthUrl(provider);
+
+    // TODO: handle error
+    if (res.code === HTTPSTATUS.success) {
+      const oAuthPopUpWindow = openPopUpWindow(
+        res.data.authURL,
+        `${provider}OAuth`,
+        500,
+        600
+      );
+
+      if (oAuthPopUpWindow) {
+        onOpenPopUpWindow(oAuthPopUpWindow);
+      } else {
+        toast.error("소셜 로그인을 위해 팝업을 허용해주세요");
+      }
+    }
+  };
+
+  if (provider === "google") {
+    return (
+      <StyledGoogleSignInButton onClick={onSignIn}>
+        <img src={googleLogo} alt="구글 로고" />
+        구글 로그인
+      </StyledGoogleSignInButton>
+    );
+  }
+
+  if (provider === "kakao") {
+    return (
+      <StyledKakaoSignInButton type="button" onClick={onSignIn}>
+        <img src={kakaoLogo} alt="카카오 로고" />
+        <p>카카오 로그인</p>
+      </StyledKakaoSignInButton>
+    );
+  }
+
+  if (provider === "naver") {
+    return (
+      <StyledNaverSignInButton type="button" onClick={onSignIn}>
+        <img src={naverLogo} alt="네이버 로고" />
+        <p>네이버 로그인</p>
+      </StyledNaverSignInButton>
+    );
+  }
+}
+
+const StyledKakaoSignInButton = styled(BaseSignInButton)`
+  background-color: ${designSystem.color.kakao.primary};
+  color: ${designSystem.color.neutral.black};
+  font: ${designSystem.font.button2};
+`;
+
+const StyledGoogleSignInButton = styled(BaseSignInButton)`
+  background-color: ${designSystem.color.neutral.white};
+  border: 1px solid ${designSystem.color.neutral.gray200};
+  color: ${designSystem.color.neutral.gray600};
+  font: ${designSystem.font.button2};
+`;
+
+const StyledNaverSignInButton = styled(BaseSignInButton)`
+  background-color: ${designSystem.color.naver.primary};
+  color: ${designSystem.color.neutral.white};
+  font: ${designSystem.font.button2};
+`;
